rename bosituasjonForm -> BosituasjonPageUtils
